Do refactoring.
