+++ ManagedDoom/src/Doom/Opening/DemoPlayback.cs
