+++ ManagedDoom/src/Doom/Opening/OpeningSequence.cs
