+++ ManagedDoom/src/Doom/Opening/OpeningSequenceState.cs
